minor
Former-commit-id: 1cfe6616fc24064aa1ac6f6cbfac8978fc5e9de6
Former-commit-id: 9ccc0982910f0914b0132b95ce24c79da2ffbbfd

+++ configure.py
@@ -86,7 +86,7 @@ def defineOptions():
   parser.add_argument('--nttiny_dir', default="extern/nttiny", help='specify path for `Nttiny`')
 
   # simulation
-  parser.add_argument('--precision', default='double', choices=Precision_options, help='code precision')
+  parser.add_argument('--precision', default='single', choices=Precision_options, help='code precision')
   parser.add_argument('--pgen', default="", choices=Pgen_options, help='problem generator to be used')
 
   # `Kokkos` specific
